requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
pbr>=1.6 # Apache-2.0

enum34;python_version=='2.7' or python_version=='2.6' or python_version=='3.3' # BSD
eventlet!=0.18.3,>=0.18.2 # MIT
httplib2>=0.7.5 # MIT
netaddr!=0.7.16,>=0.7.13 # BSD
tenacity>=3.1.1 # Apache-2.0
SQLAlchemy<1.1.0,>=1.0.10 # MIT
six>=1.9.0 # MIT
stevedore>=1.17.1 # Apache-2.0
neutron-lib>=0.4.0 # Apache-2.0
osc-lib>=1.2.0 # Apache-2.0
python-openstackclient>=3.3.0 # Apache-2.0
oslo.concurrency>=3.8.0 # Apache-2.0
oslo.config!=3.18.0,>=3.14.0 # Apache-2.0
oslo.db!=4.13.1,!=4.13.2,>=4.10.0 # Apache-2.0
oslo.i18n>=2.1.0 # Apache-2.0
oslo.log>=3.11.0 # Apache-2.0
oslo.serialization>=1.10.0 # Apache-2.0
oslo.service>=1.10.0 # Apache-2.0
oslo.utils>=3.17.0 # Apache-2.0
oslo.vmware>=2.11.0 # Apache-2.0
PrettyTable<0.8,>=0.7.1 # BSD
tooz>=1.28.0 # Apache-2.0
decorator>=3.4.0 # BSD
